+++ .gitignore
@@ -1,2 +1,3 @@
 dist-newstyle
 result
+flake
