+++ dapps-certification-helpers/dapps-certification-helpers.cabal
@@ -1,0 +1,73 @@
+cabal-version:      2.4
+name:               dapps-certification-helpers
+version:            0.1.0.0
+author:             Shea Levy
+maintainer:         shea.levy@iohk.io
+data-files: data/Certify.hs data/outputs.nix
+
+library
+    exposed-modules: IOHK.Certification.Actions
+    build-depends:
+      base,
+      aeson,
+      text,
+      eventuo11y ^>= 0.3.2,
+      bytestring,
+      typed-process,
+      base16,
+      time,
+      directory,
+      network-uri,
+      process,
+      unliftio-core,
+      containers,
+      filepath,
+      async,
+      exceptions,
+      vector,
+      resourcet,
+      conduit,
+      conduit-aeson,
+      dapps-certification-interface
+    hs-source-dirs:   src
+    other-modules: Paths_dapps_certification_helpers
+    autogen-modules: Paths_dapps_certification_helpers
+    default-language: Haskell2010
+
+executable generate-flake
+    build-depends:
+      base,
+      dapps-certification-helpers,
+      optparse-applicative,
+      network-uri,
+      eventuo11y,
+      directory,
+      aeson
+    main-is: Main.hs
+    hs-source-dirs: generate-flake
+    default-language: Haskell2010
+
+executable build-flake
+    build-depends:
+      base,
+      dapps-certification-helpers,
+      optparse-applicative,
+      eventuo11y
+    main-is: Main.hs
+    hs-source-dirs: build-flake
+    default-language: Haskell2010
+
+executable run-certify
+    build-depends:
+      base,
+      dapps-certification-interface,
+      dapps-certification-helpers,
+      optparse-applicative,
+      filepath,
+      conduit,
+      aeson,
+      bytestring,
+      resourcet
+    main-is: Main.hs
+    hs-source-dirs: run-certify
+    default-language: Haskell2010
