+++ dapps-certification-interface/dapps-certification-interface.cabal
@@ -3,69 +3,9 @@ name:               dapps-certification-interface
 version:            0.1.0.0
 author:             Shea Levy
 maintainer:         shea.levy@iohk.io
-data-files: data/Certify.hs data/outputs.nix
 
 library
-    exposed-modules: IOHK.Certification.Interface, IOHK.Certification.Interface.Actions
-    build-depends:
-      base,
-      aeson,
-      text,
-      eventuo11y ^>= 0.3.2,
-      bytestring,
-      typed-process,
-      base16,
-      time,
-      directory,
-      network-uri,
-      process,
-      unliftio-core,
-      containers,
-      filepath,
-      async,
-      exceptions,
-      vector,
-      resourcet,
-      conduit,
-      conduit-aeson
+    exposed-modules:  IOHK.Certification.Interface
+    build-depends:    base, aeson, text
     hs-source-dirs:   src
-    other-modules: Paths_dapps_certification_interface
-    autogen-modules: Paths_dapps_certification_interface
-    default-language: Haskell2010
-
-executable generate-flake
-    build-depends:
-      base,
-      dapps-certification-interface,
-      optparse-applicative,
-      network-uri,
-      eventuo11y,
-      directory,
-      aeson
-    main-is: Main.hs
-    hs-source-dirs: generate-flake
-    default-language: Haskell2010
-
-executable build-flake
-    build-depends:
-      base,
-      dapps-certification-interface,
-      optparse-applicative,
-      eventuo11y
-    main-is: Main.hs
-    hs-source-dirs: build-flake
-    default-language: Haskell2010
-
-executable run-certify
-    build-depends:
-      base,
-      dapps-certification-interface,
-      optparse-applicative,
-      filepath,
-      conduit,
-      aeson,
-      bytestring,
-      resourcet
-    main-is: Main.hs
-    hs-source-dirs: run-certify
     default-language: Haskell2010
