+++ nix/materialized-x86_64-linux/.plan.nix/dapps-certification-helpers.nix
@@ -1,0 +1,112 @@
+{ system
+  , compiler
+  , flags
+  , pkgs
+  , hsPkgs
+  , pkgconfPkgs
+  , errorHandler
+  , config
+  , ... }:
+  {
+    flags = {};
+    package = {
+      specVersion = "2.4";
+      identifier = {
+        name = "dapps-certification-helpers";
+        version = "0.1.0.0";
+        };
+      license = "NONE";
+      copyright = "";
+      maintainer = "shea.levy@iohk.io";
+      author = "Shea Levy";
+      homepage = "";
+      url = "";
+      synopsis = "";
+      description = "";
+      buildType = "Simple";
+      isLocal = true;
+      detailLevel = "FullDetails";
+      licenseFiles = [];
+      dataDir = ".";
+      dataFiles = [ "data/Certify.hs" "data/outputs.nix" ];
+      extraSrcFiles = [];
+      extraTmpFiles = [];
+      extraDocFiles = [];
+      };
+    components = {
+      "library" = {
+        depends = [
+          (hsPkgs."base" or (errorHandler.buildDepError "base"))
+          (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
+          (hsPkgs."text" or (errorHandler.buildDepError "text"))
+          (hsPkgs."eventuo11y" or (errorHandler.buildDepError "eventuo11y"))
+          (hsPkgs."bytestring" or (errorHandler.buildDepError "bytestring"))
+          (hsPkgs."typed-process" or (errorHandler.buildDepError "typed-process"))
+          (hsPkgs."base16" or (errorHandler.buildDepError "base16"))
+          (hsPkgs."time" or (errorHandler.buildDepError "time"))
+          (hsPkgs."directory" or (errorHandler.buildDepError "directory"))
+          (hsPkgs."network-uri" or (errorHandler.buildDepError "network-uri"))
+          (hsPkgs."process" or (errorHandler.buildDepError "process"))
+          (hsPkgs."unliftio-core" or (errorHandler.buildDepError "unliftio-core"))
+          (hsPkgs."containers" or (errorHandler.buildDepError "containers"))
+          (hsPkgs."filepath" or (errorHandler.buildDepError "filepath"))
+          (hsPkgs."async" or (errorHandler.buildDepError "async"))
+          (hsPkgs."exceptions" or (errorHandler.buildDepError "exceptions"))
+          (hsPkgs."vector" or (errorHandler.buildDepError "vector"))
+          (hsPkgs."resourcet" or (errorHandler.buildDepError "resourcet"))
+          (hsPkgs."conduit" or (errorHandler.buildDepError "conduit"))
+          (hsPkgs."conduit-aeson" or (errorHandler.buildDepError "conduit-aeson"))
+          (hsPkgs."dapps-certification-interface" or (errorHandler.buildDepError "dapps-certification-interface"))
+          ];
+        buildable = true;
+        modules = [
+          "Paths_dapps_certification_helpers"
+          "IOHK/Certification/Actions"
+          ];
+        hsSourceDirs = [ "src" ];
+        };
+      exes = {
+        "generate-flake" = {
+          depends = [
+            (hsPkgs."base" or (errorHandler.buildDepError "base"))
+            (hsPkgs."dapps-certification-helpers" or (errorHandler.buildDepError "dapps-certification-helpers"))
+            (hsPkgs."optparse-applicative" or (errorHandler.buildDepError "optparse-applicative"))
+            (hsPkgs."network-uri" or (errorHandler.buildDepError "network-uri"))
+            (hsPkgs."eventuo11y" or (errorHandler.buildDepError "eventuo11y"))
+            (hsPkgs."directory" or (errorHandler.buildDepError "directory"))
+            (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
+            ];
+          buildable = true;
+          hsSourceDirs = [ "generate-flake" ];
+          mainPath = [ "Main.hs" ];
+          };
+        "build-flake" = {
+          depends = [
+            (hsPkgs."base" or (errorHandler.buildDepError "base"))
+            (hsPkgs."dapps-certification-helpers" or (errorHandler.buildDepError "dapps-certification-helpers"))
+            (hsPkgs."optparse-applicative" or (errorHandler.buildDepError "optparse-applicative"))
+            (hsPkgs."eventuo11y" or (errorHandler.buildDepError "eventuo11y"))
+            ];
+          buildable = true;
+          hsSourceDirs = [ "build-flake" ];
+          mainPath = [ "Main.hs" ];
+          };
+        "run-certify" = {
+          depends = [
+            (hsPkgs."base" or (errorHandler.buildDepError "base"))
+            (hsPkgs."dapps-certification-interface" or (errorHandler.buildDepError "dapps-certification-interface"))
+            (hsPkgs."dapps-certification-helpers" or (errorHandler.buildDepError "dapps-certification-helpers"))
+            (hsPkgs."optparse-applicative" or (errorHandler.buildDepError "optparse-applicative"))
+            (hsPkgs."filepath" or (errorHandler.buildDepError "filepath"))
+            (hsPkgs."conduit" or (errorHandler.buildDepError "conduit"))
+            (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
+            (hsPkgs."bytestring" or (errorHandler.buildDepError "bytestring"))
+            (hsPkgs."resourcet" or (errorHandler.buildDepError "resourcet"))
+            ];
+          buildable = true;
+          hsSourceDirs = [ "run-certify" ];
+          mainPath = [ "Main.hs" ];
+          };
+        };
+      };
+    } // rec { src = (pkgs.lib).mkDefault ../dapps-certification-helpers; }
